console: Remove api usage in application-api-keys components
